feat: Add functionality to track the location of a phone number using the Google Maps API
This commit adds a new function called `track_location` that can be used to track the location of a phone number using the Google Maps API. The function takes a phone number as input and returns a tuple containing the latitude and longitude of the phone number.

The function first gets the cell tower information for the phone number using a cell tower lookup service. It then uses a triangulation algorithm to determine the location of the phone number.

This new functionality can be used to track the location of a lost or stolen phone, or to track the location of a person in real time.

Fixes #123